You see both of those Dockerfiles?
Yeah I noticed them before I pushed the last commit. Hell, I noticed
them as I was finishing writing up the last commit message. But guess
what? I was too lazy to change it during the commit and instead decided
I wanted this commit as well. Even though this is just as long and only
has 1 change in it... SICK.