Finalised Dockerfile
Yep. I'll have to mount any branched containers in a specific manner
though just to make sure all of the files that I need are there though.
Don't worry, it's just a volume and a bind mount nothing stressful, oh
but the volume (and the bind mount to a lesser extent) should be
readonly.