I'm THIS close to finishing the AE
But there's an absolutely fantastic dimension bug. Like it expects the
input to be 3D, when it's only 2D. But I can't make the ConvNet 2D
because it has to be 3D, even though it's a 1D convolution. Like, what?
You can't be serious. What do you expect me to do?