Finished the AE
I'm REALLY skeptical on the practicality of this. The loss converges way
too quickly. Like after the 2nd epoch, quick. I still need to jam it
into the SVM though, so it may be great or a bigger mess than...
something that's a mess. IDK I coulda made an easy Trump Organisation
joke but I'm better than that.