Some big boi changes here
So there's the process file which makes it way easier to filter posts
and comments as well as run feature extraction on a given dataset. It
needs some more work, so that you just need to supply a dataset, instead
of a dataset and a file as well. Minute changes in stack and xTract as
well. All of this just so I can easily hook up my ConvNet to my SVM...
smh. And I haven't even gotten a Random Forest going yet. SICK.