Partial CSV Generation
So CSV files are written with each model being evaluated. The CSV
allowing for a correlation between accuracy and equivalence class size
can be generated as a result of some of these changes. Post-processing
is needed on that CSV to allow for the correct plots to be generated
though, which should be fine but it's 2:18:11 and I want to sleep.